🧹 Move files from common to app
